04-FaceRecognition-II/thetensorclan-backend-heroku/requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.7.0%2Bcpu-cp37-cp37m-linux_x86_64.whl
Flask==1.1.2
Pillow==7.2.0
Flask-Cors==3.0.8
gdown==3.12.2
gunicorn==20.0.4
